% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXY_auto_install.R
\name{set_custom_mirrors}
\alias{set_custom_mirrors}
\title{Configure custom CRAN and Bioconductor mirrors}
\usage{
set_custom_mirrors(
  cran_mirror = "https://cran.r-project.org/",
  bioc_mirror = "https://bioconductor.org/"
)
}
\arguments{
\item{cran_mirror}{Character, CRAN mirror address (default: Westlake University)}

\item{bioc_mirror}{Character, Bioconductor mirror address (default: Westlake University)}
}
\description{
Automatically set specified CRAN and Bioconductor mirror addresses
}
