% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SXY_auto_install.R
\name{SXY_install_package_auto}
\alias{SXY_install_package_auto}
\title{Auto install R packages by source (SXY version, CRAN first)}
\usage{
SXY_install_package_auto(
  pkg_name,
  cran_mirror = "http://mirrors.westlake.edu.cn/CRAN",
  bioc_mirror = "http://mirrors.westlake.edu.cn/bioconductor"
)
}
\arguments{
\item{pkg_name}{Character, package name (pure name or author/package format for GitHub)}

\item{cran_mirror}{Character, custom CRAN mirror address (default: Westlake University)}

\item{bioc_mirror}{Character, custom Bioconductor mirror address (default: Westlake University)}
}
\description{
Automatically identify package source (GitHub > CRAN > Bioconductor) and install, 
with all dependencies and no auto-update. CRAN is checked first, then Bioconductor.
}
